+++ src/playground/widget-editor/src/components/editor-form/component.js
@@ -1,107 +1,29 @@
-import React, { useState, useEffect } from 'react';
-import { useSelector, useDispatch } from 'react-redux'
+import React, { useState } from 'react';
 import Select from 'react-select';
 import CreatableSelect from 'react-select/creatable';
 
+import useEditorForm from './useEditorForm';
+
 import './style.scss';
 
-const getAllWidgetsForDataset = async datasetId => {
-  try {
-    const response = await fetch(`https://api.resourcewatch.org/v1/dataset/${datasetId}/widget`);
-    const { data } = await response.json();
-    return data.map(d => ({
-      label: d.attributes.name,
-      value: d.id
-    }));
-  } catch (e) {
-     console.error('(Playground) failed to fetch widgets')
-  }
-}
-
 const EditorForm = () => {
-  const dispatch = useDispatch();
   const [active, setActive] = useState(false);
   const [autoFillValue, setAutoFillValue] = useState('');
-  const [autoFillError, hasAutoFillError] = useState(false);
 
-  const datasets = useSelector(state => state.editorOptions.datasets);
-  const widgets = useSelector(state => state.editorOptions.widgets);
-
-  const dataset = useSelector(state => {
-    const { editorOptions: { datasets, dataset } } = state;
-    const selected = datasets.find(d => d.value === dataset);
-    return selected || '';
-  });
-
-  const widget = useSelector(state => {
-    const { editorOptions: { widgets, widget } } = state;
-    if (!widget) {
-      return '';
-    }
-    const selected = widgets.find(d => d.value === widget);
-    return selected || '';
-  });
-
-  const isCustomDataset = !datasets.find(d => d.value === dataset.value);
-
-  useEffect(() => {
-    const patch = async () => {
-      const datasetId = datasets[0].value;
-      const allWidgets = await getAllWidgetsForDataset(datasetId);
-      dispatch(modifyOptions({ dataset: datasets[0].value, widgets: allWidgets }));
-    }
-    if (!dataset) {
-      patch();
-    }
-  }, []); // eslint-disable-line
-
-  const modifyOptions = payload => ({
-    type: 'PLAYGROUND/modifyOptions',
-    payload
-  });
+  const [
+    datasets,
+    widgets,
+    dataset,
+    widget,
+    isCustomDataset,
+    autoFillError,
+    handleChangeDataset,
+    handleChangeWidget
+  ] = useEditorForm(autoFillValue);
 
   const autoFill = async e => {
     const { value } = e.target;
-    try {
-      const response = await fetch(value);
-      const { data } = await response.json();
-      if (data.type === 'widget') {
-        const allWidgets = await getAllWidgetsForDataset(data.attributes.dataset);
-        dispatch(modifyOptions({
-          dataset: data.attributes.dataset,
-          widget: data.id,
-          widgets: allWidgets
-        }))
-      }
-      if (data.type === 'dataset') {
-        const allWidgets = await getAllWidgetsForDataset(data.id);
-        dispatch(modifyOptions({
-          dataset: data.id,
-          widget: null,
-          widgets: allWidgets
-        }))
-      }
-      hasAutoFillError(false);
-    } catch (e) {
-      console.error('(Playground) failed to fetch widgets')
-      hasAutoFillError(true);
-    }
-
     setAutoFillValue(value);
-  }
-
-  const handleChangeDataset = async (item, { action }) => {
-    const allWidgets = await getAllWidgetsForDataset(item.value);
-    dispatch(modifyOptions({
-      dataset: item.value,
-      widgets: allWidgets
-    }))
-  }
-
-  const handleChangeWidget = async (item, { action }) => {
-    dispatch(modifyOptions({
-      widget: item.value
-    }))
   }
 
   return (
